+++ MadmucFarm/table/Field.cs
@@ -11,7 +11,7 @@ namespace MadmucFarm
 		[PrimaryKey,AutoIncrement]
 		public int fieldID{ get; set; }
 		public string fieldName{ get; set;}
-		public int acre{ get; set; }
+		public double acre{ get; set; }
 		public int farmID{ get; set;}
 		public string note{ get; set;}
 	}
